- Navnendringer og opprydding (fjernet ubrukte filer, renamet osv )
