--- src/api/domain/GuiModelMetadata.ts
@@ -1,8 +1,0 @@
-interface GuiModelMetadata {
-    status?: string,
-    info?: string,
-    bruktSats?: string,
-    debugLog?: string,
-}
-
-export default GuiModelMetadata
