+++ src/api/domain/types/guimodel.ts
